chore: optimize hooks
