+++ packages/shared/hooks/useMutationObservable/index.md
@@ -1,0 +1,1 @@
+useMutationObservable
