+++ packages/shared/hooks/useVariables/index.md
@@ -1,0 +1,1 @@
+useVariables
